playing with vuestic UI
