+++ candle/src/App.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 // This starter template is using Vue 3 <script setup> SFCs
 // Check out https://v3.vuejs.org/api/sfc-script-setup.html#sfc-script-setup
-import HelloVuesticUi from "./components/HelloVuesticUi.vue";
+import PlayersOverview from "./components/PlayersOverview.vue";
 </script>
 
 <template>
-  <HelloVuesticUi />
+  <PlayersOverview />
 </template>
 
 <style>
